requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
langchain-text-splitters
sentence_transformers
langchain_huggingface
pandas
langchain-groq
youtube_transcript_api
unstructured
pytube
langchain-yt-dlp
numexpr
huggingface_hub
ffmpeg
imageio[ffmpeg]
datasets[audio]
accelerate
validators
